refactor(http): move application
