+++ lib/nonnative.rb
@@ -23,6 +23,7 @@ require 'nonnative/command'
 require 'nonnative/pool'
 require 'nonnative/server'
 require 'nonnative/http_client'
+require 'nonnative/application'
 require 'nonnative/http_server'
 require 'nonnative/grpc_server'
 require 'nonnative/grpc_server'
